+++ pidman/admin.py
@@ -1,9 +1,12 @@
 from django.contrib import auth
 from django.contrib.admin import AdminSite
 from django.contrib.admin.models import LogEntry
+from downtime.models import Period
+from eultheme.models import Banner
 from logentry_admin.admin import LogEntryAdmin
 
 import sequences
+
 
 class PidmanAdminSite(AdminSite):
     site_header = 'Persistent Identifier Manager'
@@ -17,3 +20,16 @@ admin_site.register(auth.models.Group, auth.admin.GroupAdmin)
 admin_site.register(auth.models.User, auth.admin.UserAdmin)
 admin_site.register(sequences.models.Sequence, sequences.admin.Sequence)
 admin_site.register(LogEntry, LogEntryAdmin)
+# eultheme / downtime
+admin_site.register(Period)
+admin_site.register(Banner)
+admin_site.register(LogEntry, LogEntryAdmin)
+
+
+def template_settings(request):
+    '''Template context processor to add settings for use on any page.'''
+
+    context_extras = {
+        'site_title': PidmanAdminSite.site_header
+    }
+    return context_extras