+++ pidman/pid/__init__.py
@@ -1,0 +1,3 @@
+'''Core application to represent pid objects and build central behaviors
+for them.  All (non-scripted) management and editing is handled
+through the Django Admin interface.'''