+++ pidman/pid/admin.py
@@ -1,11 +1,13 @@
-from pidman.pid.models import ExtSystem, Pid, Proxy, Target, Policy, Domain
 from django.contrib import admin
 from django import forms
 from django.forms.models import ModelForm, ModelChoiceField
 from django.forms import ValidationError
 
+from mptt.admin import MPTTModelAdmin, MPTTAdminForm
+from mptt.fields import TreeNodeChoiceField
 from pidman.admin import admin_site
 from pidman.pid.ark_utils import normalize_ark, invalid_qualifier_characters
+from pidman.pid.models import ExtSystem, Pid, Proxy, Target, Policy, Domain
 
 
 class TargetInlineForm(ModelForm):
@@ -22,6 +24,7 @@ class TargetInlineForm(ModelForm):
         # normalize according to how the ARK will be resolved
         return normalize_ark(self.cleaned_data["qualify"])
 
+
 class TargetInline(admin.TabularInline):
     model = Target
     fields = ('qualify', 'uri', 'proxy', 'active')
@@ -32,37 +35,20 @@ class TargetInline(admin.TabularInline):
 # NOTE: should be possible to extend inline template here
 # to display link status - last checked / message / etc
 
+
 class PurlTargetInline(TargetInline):
     verbose_name_plural = "Target"
     max_num = 1
     can_delete = False      # do not allow PURL target deletion (only one target)
     fields = ('uri', 'proxy', 'active')
 
-# customized domain-specific choice field - sets choices differently
-class DomainChoiceField(ModelChoiceField):
-    # generate a list of domains with grouped collections for building select
-    cached = None
-
-    @property
-    def _choices(self):
-        # If a ModelChoiceField has a '_choices' attribute, it is used instead
-        # of the model queryset.
-         for d in Domain.objects.filter(parent=None):
-            yield (d.id, d.name)
-            if d.collections.count():
-                # If a top-level domain has collections, add a label and
-                # generate a list of those collections.
-                # This is interpreted by the select wigdet as an option group
-                # with a label and choices, so that collections are clearly
-                # labeled by the domain that they belong to.
-                yield (d.name + ' collections',
-                        [(c.id, c.name) for c in d.collections.all()])
 
 class PidAdminForm(ModelForm):
-    domain = DomainChoiceField(queryset=Domain.objects, cache_choices=True)
+    domain = TreeNodeChoiceField(queryset=Domain.objects.all())
     class Meta:
         model = Pid
         exclude = []
+
 
 class PidAdmin(admin.ModelAdmin):
     # browse display: type (ark/purl), domain/collection, name/description, and pid url (not target url)
@@ -70,14 +56,18 @@ class PidAdmin(admin.ModelAdmin):
     # including dates in list display for sorting purposes
     # sort columns by: type, domain/collection, name, (pid url?), date created/modified ascending/descending
     list_display = ('pid', 'truncated_name', 'type', 'created_at', 'updated_at',
-        "domain", "primary_target_uri", "is_active")  #, 'linkcheck_status')
+                    'domain', "primary_target_uri", "is_active", 'linkcheck_status')
     # filters: collection/domain, creator/user, type (ark/purl), date ranges (created or modified)
-    list_filter = ['type', 'domain', 'ext_system', 'creator', 'created_at',
-        'updated_at']
+    list_filter = (
+        'type', ('domain', admin.RelatedOnlyFieldListFilter),
+        'ext_system',
+        ('creator', admin.RelatedOnlyFieldListFilter),
+        'created_at', 'updated_at')
+    list_select_related = True  # may want to limit this some
     form = PidAdminForm
 
     # now possible in django 1.1 - fields to use here?
-    #list_editable = ('name', 'domain')
+    # list_editable = ('name', 'domain')
     date_hierarchy = 'created_at'
     search_fields = ['name', 'pid', 'ext_system_key', 'target__uri']
     # keep pid type in a separate fieldset in order to suppress it on edit
@@ -116,24 +106,49 @@ class PidAdmin(admin.ModelAdmin):
             obj.creator = request.user
         obj.save()
 
-    #disallow delete of Pids set targets to inactive instead
+    # disallow delete of Pids set targets to inactive instead
     def has_delete_permission(self, request, obj=None):
         return False
+
+    def get_queryset(self, request):
+        # extend queryset to prefetch targets & linkcheck status,
+        # which are used in the change list display
+        pidqs = super(PidAdmin, self).get_queryset(request)
+        pidqs = pidqs.prefetch_related('target_set', 'domain', 'creator',
+                                       'target_set__linkcheck__url')
+        return pidqs
+
+    def formfield_for_dbfield(self, db_field, **kwargs):
+        request = kwargs['request']
+        formfield = super(PidAdmin, self).formfield_for_dbfield(db_field, **kwargs)
+
+        if db_field.name == 'domain':
+            choices = getattr(request, '_domain_choices_cache', None)
+            if choices is None:
+                request._domain_choices_cache = choices = list(formfield.choices)
+            formfield.choices = choices
+
+        return formfield
 
 
 class ExtSystemAdmin(admin.ModelAdmin):
     list_display = ('name', 'key_field', 'updated_at')
 
+
 class ProxyAdmin(admin.ModelAdmin):
     list_display = ('name', 'transform', 'updated_at')
+
 
 class PolicyAdmin(admin.ModelAdmin):
     list_display = ('commitment', 'created_at')
 
-class DomainAdminForm(forms.ModelForm):
-    # restrict list of domains allowed to be parents to those domains without a parent (1 level deep only)
-    # FIXME: is there any way to exclude current domain from this list? no access to instance id here
-    parent = ModelChoiceField(queryset=Domain.objects.filter(parent=None).all(), required=False)
+
+# class DomainAdminForm(forms.ModelForm):
+class DomainAdminForm(MPTTAdminForm):
+    # restrict list of domains allowed to be parents to those domains
+    # without a parent (1 level deep only)
+    parent = TreeNodeChoiceField(queryset=Domain.objects.filter(parent__isnull=True),
+        required=False)
     class Meta:
         model = Domain
         exclude = []
@@ -143,34 +158,44 @@ class DomainAdminForm(forms.ModelForm):
         if parent:
             # check parent id - cannot point to self
             if parent.id == self.instance.id:
-               raise ValidationError("Not permitted: a domain can not be its own parent");
+                raise ValidationError("Not permitted: a domain can not be its own parent",
+                   code='invalid')
             # restrict hierarchy to one level
             elif parent.parent:
                 raise ValidationError("Domain hierarchy restricted to depth of 1; " +
-                                parent.name + " is a collection of " + parent.parent.name)
+                                parent.name + " is a collection of " + parent.parent.name,
+                                code='invalid')
         return parent
 
     def clean(self):
-        #raise ValidationError(self.cleaned_data)
         # policy is optional by default, but top-level domains must have one (can't inherit from parent)
-        if not self.cleaned_data['parent'] and not self.cleaned_data['policy']:
-           raise ValidationError("Policy is required for top-level domains");
+        if not self.cleaned_data.get('parent', None) and not self.cleaned_data['policy']:
+           raise ValidationError("Policy is required for top-level domains")
         return self.cleaned_data
+
 
 class CollectionInline(admin.TabularInline):
     model = Domain
-    verbose_name = "Collection"
+    verbose_name = "Domain"
     verbose_name_plural = "Collections"
+    # parent = TreeNodeChoiceField(queryset=Domain.objects.all(),
+        # level_indicator=u'+--')
 
-class DomainAdmin(admin.ModelAdmin):
+
+class DomainAdmin(MPTTModelAdmin):
     form = DomainAdminForm
-    list_display = ('name', 'num_collections', 'num_pids')
+    mptt_level_indent = 20
+
+    list_display = ('name', 'num_pids', 'subdomain_count', 'show_policy')
     inlines = [CollectionInline]
 
-    # extend queryset to limit list view to top-level domains
-    def queryset(self, request):
-        qs = super(DomainAdmin, self).queryset(request)
-        return qs.filter(parent=None)
+    def get_queryset(self, request):
+        # extend queryset to prefetch policy and parent,
+        # which are used in the change list display
+        domains = super(DomainAdmin, self).get_queryset(request)
+        domains = domains.prefetch_related('policy', 'parent')
+        return domains
+
 
 admin_site.register(Pid, PidAdmin)
 admin_site.register(Proxy, ProxyAdmin)