+++ pidman/urls.py
@@ -5,10 +5,13 @@ from django.views.generic.base import RedirectView
 
 from pidman.admin import admin_site
 
+
 urlpatterns = [
     # pidman has no user-facing index, so redirect from base site url
     # to site admin page
-    url(r'^$', RedirectView.as_view(url='admin/', permanent=False)),
+
+    url(r'^$', RedirectView.as_view(url='admin/', permanent=False),
+        name="site-index"),
     url(r'^admin/linkcheck/', include('linkcheck.urls', namespace='linkcheck')),
     url(r'^admin/', include(admin_site.urls)),